Add files via upload
The presence of monkeys at temples and educational institutions is creating challenges by disrupting daily activities. Traditional methods involving human intervention have proven to be ineffective and inefficient. To address this issue, a modern approach is proposed. It involves using IP-based cameras equipped with sensors and AI algorithms to detect monkeys. When the system detects monkeys, it activates a mechanism to deploy unconscious injections, tranquilizing them safely. The process includes precise targeting, communication with a control center for monitoring and response, and eventual removal of tranquilized monkeys by personnel. This method aims to mitigate disruptions caused by monkeys while ensuring their humane treatment and the safety of the affected areas.